requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.1
cloudinary==1.38.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
